Add a Simple array to sanitize operation.
This also added transfer of the parameterDefinition through the
processCustom function to allow the custom function to know the associated
definition.

Added a check within the function
filterStrictSanitizeValues to apply/allow assignment of strict
sanitization to a field and for it to be strictly sanitized if the
overall process supports it. This means that if a field is specifically
identified for strict sanitization and the system is set to not do strict
sanitization, then the field will not be strictly sanitized.

    Address STRICT_SANITIZE_KEYS and duplicate parameters

    Added action and checks to remove keys that should be sanitized out of
 the operation.

This takes the main parameter and merges it with the sub-parameter
to generate a unique key for the sub-parameter by the addition of the
internal function setCurrentArrayName($parameter, &$currentArrayName = NULL)
which can be used to add the key to the existing array name to support
generating a unique identifier for the content posted to be processed.

Complete full sanitization for undefined params

If a sanitizerType is identified as SIMPLE_ARRAY but there is no further
definition for 'params' then this modification should ensure sanitization
of all sub-keys.

Add in feature to track which item has specifically been processed.

In order to know to apply sanitization, the specific post key was needed
to be known and this adds that feature so that in the sample code that
initiated the development of this feature the key name could be reused.

Sample code was something like:
file[id]
file[url]
fields[id]
fields[url]

the presence of id and url in both file and field requires the ability to
know that the file and fields keys are different and that they may require
different sanitization patterns.