Reduce db queries for prices (#1473)
Reduce db queries for prices

In several of the price functions, both the special and sales price is sought.
Both are obtained through the same function `zen_get_products_special_price` with a switch to either return only the specials price (true) or to return the overall special price (false) (sale price if present otherwise special price if a special, and false if neither)
If a product is not on special and does not have a sale then the returning result is `false` and therefore there is no further determination of the special necessary (ie. no further database query needed). If the result is anything other than `false`, then there may be a special and the value of that special (if present) remains to be determined.

While there remain other means of gaining efficiency in this process (such as returning all data collected from a one time database search/review), this at least reduces the redundancy of querying the database for results of the first `zen_get_products_special_price` call.  Though perhaps caching also helps in this overall concept, but the recommended modification mitigates the need to use the cache.