Flush out return to checkout_shipping if shipping method not selected.
The checkout_shipping process is expected to support moving on through
the checkout process if a shipping method is identified. If
$_SESSION['shipping'] is set to false through a call to $shipping->cheapest()
where the only available shipping method is instore (or whatever else has
been set to be ignored in that class, then a shipping method will not be
forced if the URI is then entered as checkout_payment.  The same is true
in the current configuration if all shipping methods are disabled or there
are no shipping methods defined.  (That it be possible to continue through
checkout without ever having selected a shipping method nor one being
available.)