v156: 0-weight/freeshipping message might not appear
The message "Warning: 0 weight is configured for Free Shipping and Free Shipping Module is not enabled" intended to warn the admin that the setting _Configuration->Shipping/Packaging->Order Free Shipping 0 Weight Status_ is enabled but the `freeshipper` shipping module is not enabled.

Unfortunately, the message does **not** currently appear if `freeshipper` is installed but disabled.  This PR makes that correction in both the admin's configuration.php and modules.php scripts.