v156 rework query_factory class
Multiple tasks accomplished here:
1. Primary goal: review of the values from a foreach database query will
no longer maintain an immutable series of numeric keys that correspond to
the result of the first record of a database query.
2. Secondary goal: reduce the memory usage of queries.
3. Maintain existing functionality in base method use, adding features
that allow additional control to better support the above.
4. Support returning all associative field data not just those that do not
begin with a number. (Note that when referencing field data turned to an
object that begins with a number requires a different format of accessing.
ie. if a field's name were 3e then it was converted to an object it
could be accessed via: $this->{'3e'} )