v156-notifiers-for-base-functions
Add notifiers to various "base" functions, enabling observers to modify (or in some cases replace) those functions' processing and return values.  This will enable many plugins to remove the need for core-file overwrites, allowing for an easier Zen Cart upgrade path.

Please note that many of the notifier strings are in use by various plugins; please don't change the names and/or "payload".