fix #1736
With the new edited categories/products code somewhere before or near
the commit: 2e7ddad
it appears that if have navigated to a category then selected to add a
product, the product ends up being added to the root of the store rather
than the current category. It must then be moved to the desired category
to place it where desired.

Historically, once a user has navigated to a category and then selects
to add a product, the category at that point becomes the master category
for the product. That does not appear to be happening now.