Prepare for filter data.
While resolving strict notifications identified the use or existence
of the variable $do_filter_list in the advanced_search_result page.

This is defined only in the index_filters section, but also not consistently
so either $do_filter_list should be removed or more appropriately captured.

Chose to pull the two together so that whatever functionality someone
may have been using it would still be there, but at least the strict
messages would not be generated.