Revise product type check
There is duplication in the code to identify the product types that can be removed by reworking some of the variables and operations.  The effect of this modification is that there is no longer a variable created called `$product_types_array` that would only exist if there were no product restrictions, there is no longer a specific variable generated `$restrict_types` to identify the specific restricted product types (or absence of them).

Also applied the PHP 5.6+ use of `[]` for array declaration instead of `array()` throughout.