Direct product entry to product type preview and correct preview modification.
In #1736, the admin/product.php file was modified to assign the product
type based on first the url containing an assigned products_id (pID), and if
that is not present, then to pick up on a posted value (product_type).

Then in #1747, the problem identified was that it was not possible to
transition from a non-basic product type to the preview of information.
The solution proposed supported previewing information; however, because of
the above logic and posting the data instead of clicking a link, the
product type defaulted to a standard product in admin/product.php.

This works for product that have already been assigned/created and for the
first click from the categories_to_products listing to create a new product.
However, a product's id is not assigned until the product is finally
inserted.  The rewrite also modified the program flow which includes the
central focus of the admin/product.php file needing to be provided information
to be able to know/determine the product_type (and other necessary
information), which can be accomplished a few other ways as well, but the
majority of the process flow is by way of posted data instead of url related
data.

This commit adds a hidden field to the collect info page to capture the
product_type expected, removes the product_type designation from the
uri for all default Zen Cart product types, and modifies the logic to first
act on the posting of data instead of information on the uri.

Fixes #1773