Restore call to Move in method rewind
In #1633, there had been some ideas proposed, one of which was to
add a parameter to the call to the Move method that would specifically
rewind the query results and not offer the integer based field keys.

The functionality within the Move method was basically returned to its
previous state and the additional parameter was identified as unnecessary.

This removes the parameter and associated comment for "cleanliness".