Support adding debug backtrace data to DB Transaction log
While attempting to identify the cause, path, and history of queries
in a plugin, I found it necessary to identify the source of the
query(ies) beyond what is provided by the default logging.  At one point
I had corrected a relatively obvious source, but in so doing I lost any
additional markers of the flow path.  Adding the backtrace information
to this log (location within output is welcome to be moved as necessary)
greatly added useful information.

While such logs are already lengthy, this has been incorporated at the
moment as a feature that can be enabled by adding a constant and defining
it to not be falsey.  If this feature is considered accepted then I think
that in the logging section of the configuration menu, it would be worth
adding an additional option to include this data instead of "requiring"
a file modification.