Check query result for EOF
Back in what looks like #92, a statement was added to
admin/includes/functions/general.php within the zen_products_lookup
function to test if the database query resulted in no results by
checking for the End Of File (EOF) flag being set true.

It appears though that the wrong variable was tested.  The
current value tested is the lookup_field's result which has not
yet been processed to give an EOF style answer.  It would
seem that instead the product_lookup query result should be used
and if that result is at the EOF, then to return an empty quote,
otherwise return the desired field.

That said, the visual logic could be improved such that if there are no
values to be returned then there is no need to assign $return_field to
anything and that it could be moved outside.

Incorporated first round of comments and continued to modify function to
incorporate other changes that have become routinely applied:
- remove quotes around integer value query fields.
- captilize mysql commands