Downloads: Redirect to timeout using SSL
As @mc12345678 so kindly pointed out on the OPC repo (https://github.com/lat9/one_page_checkout/issues/151), there's a discrepancy between the `includes/modules/pages/downloads/header_php.php`.

That is due to the fact that the zc156 repo's version of the module hasn't been updated to provide that handling, which this PR corrects.