fix more info link on salemaker (#1888)
The more info link (displayed when a catrgory is alread contained in other salemake sales)
was enclosed inside the checkbox class. This meant clicking on the more info link would toggle the checkbox state
have move it to its own column. While the css may not be perfect, think this is good enough just for release