v156 post before get on cPath assignment
Generally speaking, it seems that when given the opportunity to read
something from the uri and from the posted page data, that the information
to keep/capture should come and normally comes from the posted page data.

As such, this commit modifies the sequence of testing to assign the `$cPath`
on page load such that a modification to the page's uri would not affect
the cPath.