Propose Update on admin css tables borders +others
Bootstrap collapses the tables borders, however they are reset on zencart css
In my opinion, they look better with collapsed borders. 

+others :
TH elements already are set to bold, so font-weight bold for dataheadings could be removed from the css file. 

The H1 is set to be 16px by zencart css, however H2 is set to 30px by bootstrap. Visually is not pretty.
Of course none of this is actually urgent but, I guess with time it could be altered. 
Or not.