v156 admin/ezpages.php html validation of added class parameter.
Have seen this issue in other admin screens where an input field
is drawn which has (or doesn't have) some data concatenated with
a designated class.  The text concatenated does not include a
space between the closing of the previous parameter and the start
of the next parameter.

Fixed that here for admin/ezpages.php.