v156 admin/specials.php strict notice $_GET['manual'] not set
Attempting to cast as `int` a non-existent $_GET parameter caused a
notice to be logged.  This uses the empty to circumvent the
casting if the value is set to a falsey value; however, still
kept the `int` cast because if the expectation was that `'manual'`
could be text and the desire was to evaluate that when cast to
an integer it is 0, then take a particular action.  I am not
familiar with that parameter so did not want to remove it and
its evaluation without further discussion.