v156 admin/whos_online.php rework session_decoding
When detecting the presence of a session variable, the existing
code will set the start position to 0 instead of the returned
value of false.  What this further appears to mean is that the
session record to be processed will begin at the start of the
session_data and go until the first semi-colon is encountered.

This will not process the associated session field if the field
can not be found in the provided session_data and therefore not
potentially process stored session data that doesn't relate to
the sought session record.