Add additional reporting for multi-line queries
It seems that in the "code" that processes the update sql that there is some benefit in knowing the status executing the sql queries.  When a query is written over multiple lines, log reporting does not occur if the processing code is not told to expect multiple lines.

This adds a multiple line notification to all queries that extend over multiple lines.