Cast the year in zen_date_short to int
Recently a forum member identified a problem with ZC 1.5.6 where an error was thrown I think on line 208 indicating that parameter 6 was expected to be an integer; however, a string was provided.  Now that I have had a look at the software on a larger screen, it appears that the issue stems from `$year` not being cast as an integer.