Return false from date functions evaluated to 0000-00-00 00:00:00
I have been trying to track down how a `date` field in a plugin ended up being stored with `0000-00-00` and at least the last place where assignment was made, a Zen Cart date function was used in a similar manner as other `date`/`datetime` functions.  Looking into the associated admin date functions, it appears that if the `datetime` of `0001-01-01 00:00:00` or an empty string (possibly by loose comparison) was fed into the functions then false is returned; however, there is no similar review/response if the result were `0000-00-00 00:00:00` which at least for `date` and `datetime` fields is considered to not be acceptable.

Perhaps a similar review should be done for a year of 1, month of 1, day of 1 and empty for the other three fields. Reason? a date field may have `0001-01-01` but will not include the time portion and therefore will evaluate back in the `DATE_TIME_FORMAT` as equivalent to `0001-01-01 00:00:00`... Just some thoughts

I would not be disappointed if the reason this is not incorporated is that perhaps there is a non-database reason to actually return `0000-00-00 00:00:00` and therefore not be incorporated/included.