v156 correct display of attribute price for priced-by-attribute
Identified in https://www.zen-cart.com/showthread.php?224923-Attribute-pricing-display-all-fouled-up, if a product is `priced by attribute`, the default resulting price included the `products_price`.  Removing the `products_price` from the attribute price response restores the previous default functionality; however, doing that for all situations causes the result of the name of this function (`zen_get_attributes_price_final`) to not be possible.  Added a parameter that defaults to the historical ZC response of not including the `products_price` in the final result.  This solution resolved the observed condition and may maintain the overall desired resolution made by pull #1404.  If not, then the subtraction/addition should be that of zen_get_discount_calc without the `options_values_price` factored into that price difference.