Finally correct priced-by-attributes, special, sale attribute display
Previous attempted corrections caused product priced-by-attributes with only a price_factor to display the attribute price + the attribute's price.  This corrects that issue and supports the attribute's display. It also appears to correct an issue identified in the following threads:
- https://www.zen-cart.com/showthread.php?219366-SaleMaker-only-discounts-base-price-not-attributes
- https://www.zen-cart.com/showthread.php?221940-attributes-price-factor-special-price-incorrect-totals&p=1326998#post1326998
- https://www.zen-cart.com/showthread.php?221940-attributes-price-factor-special-price-incorrect-totals

It also removes my latest error of not correctly typing: $prod_priced_by_attr in my last commit.