In modifying the core load sequence to support modifications to the sanitization script, noticed that the user notes for files loaded had not been updated to reflect the changes that have been made to file locations. Attempted to update the information to at least address class locations based on current autoload_func functionality.
Also: Remove unnecessary reset before foreach
the call to reset $autoLoadConfig in includes/autoload_func.php
was identified before a foreach.  The foreach automatically
starts from the beginning of the array, so there is no need to
make an additional call to reset it and ksort doesn't sort
based on the current pointer but by the entire dataset itself.