Address other strict notifications when searching not using the last row of search options.
A convenient debugging tool had been added that allowed the
display of lines above and below the line that contains the
result; however, if any other search tool option is chosen
from the screen, then no default set of values is identified and
in strict php processing a notification is made.  This at least
supports defaulting to a sanitized value of the $default_context_lines
when the context_lines are not otherwise provided.