Correct upload error that can occasionally occur in logs.
Found a log being generated if attempt to access a page that would
normally be the one just after an upload, but where it was
landed at without posting data.  As a result, this patch
appears to resolve the logged messages and would assign a
value if one existed.

Example to reproduce would be something like accessing:
admin/product.php?cPath=4&pID=1&action=new_product_preview