Build language dependent variables when preview_info URI entered
Previous correction applied initiated the concept of building the expected language dependent variables; however, did not incorporate the `language_id` into the variable and therefore an array was not created.  This provides the array needed to continue processing and show the information that existed about the product when it was accessed directly.