Support deleting customer_basket related info with attributes.
Identified by PanZC2020 at: https://www.zen-cart.com/showthread.php?224986-Update-SQL-error-on-customers_basket-table&p=1354071#post1354071

Inhibits update from completing because casting a string that
starts with a number appears to throw an error as described
in the referenced thread.

This supports pulling the number of the products_id if it is
stored with a colon (:) in it by retrieving everything to the left
of the colon or the entire string if a colon is not present.