Direct Details button to product's category product listing
Correct one mistake made in #1604 of the 7,700 additions and 10,077 removals.

Returns the admin user to the category list showing the product in question.  If the product is in the first page then the entire category is displayed with the product in the list, if the product is in a subsequent page, then just the product is displayed. (Separate issue?)