Correct logic mistake of pull 2093
In PR #2093, an `isset` was added to attempt to address strict notifications; however, the logic used in this area seems to be reversed in comparison to some other date related comparisons.  As a result, the added logic works in reverse as desired.  This pull sets the date to `'0001-01-01'` if the data is not posted or if it is posted and has no value in it.  If there is something of value posted, then that date will be processed.