Simplify maintenance and readability
Reduced initialization of the three array types down from an open array to one that has four elements to it.  Note an additional area of improvement here might be to use a single variable that is an array with an element for each of `spider`, `guest`, and `user` instead of maintaining separate variables for each.

By incorporating the early assignment of the variables, subsequent status assignment could be removed if the idea is that any additional status key added would also be incorporated into the initialization array.  Otherwise the check for `emptiness` when incrementing the status value should be added back in.  I have removed it here as for base code it is unnecessary.

Note also that currently at least in PHP 7.2, the display of the count without the 'Elvis operator' `($value ?: '')` would be sufficient to just display the value.  The integer zero (0) is not pushed to the display and works like the double quote; however, there is no telling what the future of PHP is to hold and whether such a translation from 0 to '' would be needed.  

What I'm saying by this is that the issue primarily addressed by this pull request branch could have been resolved simply by the initialization line that was added/modified.