Incorporated comments
Used variable $page to simplify code associated with identifying
to which page to direct work flow.
assured that $_GET['action'] had a value of some type.
Modified sanitization of data to database to cast as data type
instead of just passing along information that would otherwise
be sanitized but not to assign the specific datatype.
Restored grouping of data for comparison to ensure appropriate logic.
By eliminating an unset value for $_GET['action'] was able to
restore the switch operation to use the desired textual comparison.