156: Using pre add results in error
Funny this never was reported, but when using the pre add feature, the $specials_price variable is never set. This results in a database error because $specials_price is not a proper decimal.
Removing the variable will result in sql using the default value for the field (0.0000)., and thus continuing to the next page.