Corrects code for processing the val_function of configuration
Adds class to which additional callback type functions could
be added to supplement standard filter_var style filters and
validators.
Adds an email filter to support filtering email addresses such as
CONTACT_US_LIST.

The new class function can be access by the 'options' key
of the val_function being set to array('configurationValidation', 'sanitizeEmail')
and the 'id' set to FILTER_CALLBACK or integer equivalent.

This also modifies the admin sanitizer to bypass the
configuration_value sanitization if there is a val_function for
the associated configuration_key, and sanitizes as normal if
there is no val_function.

Changed the return on failure to support editing the value again.

Corrected call to link generator from `zen_admin_href_link` to
`zen_href_link`.

Lastly refactored the query in admin/configuration.php to remove
the additional quotes around the casting to integer.

Added a language define for the configuration page to support
the response text that might be expected for the default installed
val_function call.

Fixes #2263.

Do think that perhaps the return message could/should include
the content that was originally entered, even if it is strictly
sanitized to help the data enterer.  Haven't implemented this yet.