product_reviews_write header_php
Initially began with correction to the check of an customer being logged in (strict processing) that led to other review considerations:
require of file not to use parentheses.
Potential of the products_id not being fed to access the page.
Refactoring for code alignment.
Remove else statements where the else action is simple. (assign a default value then modify if conditions warrant an alternate result).
Add notifiers to leverage options available.