v156 Shipping Estimator retain selections
As identified in issue #2234, if:
- a user is logged in, 
- has gone beyond the `shopping_cart` page, 
- has changed the destination (`sendto`) address,
- has returned to the `shopping_cart` page, and
- the store is setup to display the potential shipping charges on the `shopping_cart` page.

Then when they arrive back at the shopping cart page, the selection determined to be active is the customer's default address that is then carried through the remainder of the checkout process until the customer again changes the selection.  This commit chooses the selection that was previously made/available.

Also, for a customer that is not logged in and because of PR #1952, if the shipping estimator is presented on the shopping cart page, the zip code is not provided and a country has been selected, then upon refresh of the page, the country selection is restored back to the store's default address.  This commit moves the strict notification handling specifically to the zip code within this section of code using the variable $zip_code that is set earlier in processing to be set to the posted zip code if provided, otherwise the existing session's zip code and if that is not present to set it to an empty string (`''`). This area is expected to be executed after the shopping_cart has been visited, but before logging in for ordering allowing the posted zip code to take the place of the existing session zip code if it exists.
Fixes #2234.