Standardizes zen_is_logged_in detectable by page.
Newer PHP versions will throw an error if a variable is referenced
that does not exist.  A user that is not logged in results in
`$_SESSION['customer_id']` not being defined.  The path forward
is to use the function zen_is_logged_in to determine if a user
is logged in.  This function supports a notifier and on a per
page basis it is possible to determine the page accessed when
the notifier fires.  This standardizes other modifications that
have been made to prevent error notification by changing
`!$_SESSION['customer_id']` and `empty($_SESSION['customer_id'])` to
just use `zen_is_logged_in`.