Admin Customers DOB not displayed on edit
When displaying a customer's record, the DOB if set to be displayed/used does not get displayed when editing the record.  This corrects that. For some reason the variable `$customers_dob` was being used which would be appropriate on update as it is set in the `update` action, but not on edit where the variable is not set.