v156 account_edit recognize ISO-8601 DOB
Similar to #1518, but for account edit.  If using a mobile device, the date is presented with a date selector (though still haven't been able to get the date selector to populate with the known/existing date) that stores the date locally/internally using ISO-8601 format which in the previous format would prevent updating the record.  This at least allows updating the record, though it seems that the user would have to re-enter the birthdate.

It appears that in jscript_responsive_framework.php the line associated with the dob:
`$('input#dob').clone().attr('type','date').insertAfter('input#dob').prev().remove();`

causes the attribute of `value` to get cleared once the type is modified to date. In briefly trying to address this in light of upcoming distribution, I haven't come up with a "one-liner" to resolve this.  May require more than one line in that file. That though is basically a separate issue. :/