v156 populate dob on initial account_edit pageload
When the `includes/modules/pages/account_edit/header_php.php` is first loaded, the value of `$dob` is not set partially because the storage validation section is bypassed.  This populates `$dob` so that when the template page is displayed, the existing database value will be available. In the existing code, once the page is posted, if there is an error to be addressed, then the posted content carries
through. Unfortunately, the original stored information was removed from being displayed.

This restores display of the stored birthday date if the template supports the display.