v156 Correct verification of email type
Although I do not see where the `customers_email_format` could be set to `NONE` or `OUT`, the variable `$email_pref_out` does not exist at
the point of the logic test.  The result is that at least in an environment with strict notification, the standard page load results in generation
of a PHP notice.