Coupon admin:  Can't add Free Shipping only coupon
When you want to create a coupon that's only valid for free shipping, you enter the coupon amount as 0 (or 0.00).  Somewhere in the process, that value gets changed to an empty string ('') which is an invalid numeric value and results in a MySQL error when the coupon is being recorded in the database.

This PR forces the amount value to be a valid numeric when that value is 'empty' (e.g. either 0, 0.00 or an empty string).