v156 Download by redirect if `HTTP_SERVER` has sub-directory.
In the following Zen Cart Forum thread:
https://www.zen-cart.com/showthread.php?225481-Doc-Product-Type-Download-by-redirect-redirects-to-wrong-pub-folder

it was identified (indirectly) that if `HTTP_SERVER` contained
a sub-directory (such as for a shared server account where the
store were reached using domain/~user, domain/user, etc...)
that when continuing on with processing, the attempted browser
path would exclude the sub-directory contained in `HTTP_SERVER`.

Unfortunately, the same would be true if the download was
expected for `HTTPS_SERVER`; however, the redirect code already
accounts for that if the provided path is based off of the
`HTTP_SERVER` prefix.

The other "unfortunate" part of this is that the web browser
path provided is assigned to a class variable which is
accessible within the attached listener for
`'NOTIFY_DOWNLOAD_READY_TO_REDIRECT'`.  The reason this is
unfortunate is that to whatever the variable is set, it is
ideal if the observer can "readily" identify what may need to be
stripped or detected if working from that direction instead of
simply building what is needed/desired to operate on the
`wsPubFolder` property.

As a result, I have created two different commits that make
up this branch, each providing the same end result of supporting
a download via redirect for a site that has one or more
sub-directories in `HTTP_SERVER`.  I am not familiar with the
other uses and/or applications that anticipate using the
above notifier and therefore welcome comment/discussion about
the best way forwards.  This first commit is the easiest and
one that seems to make the most sense for supporting further
processing and alignment to an outside storage location. E.g.
with the path built from `HTTP_SERVER` and `DIR_WS_DOWNLOAD_PUBLIC`
it is possible to just remove the beginning of the string where
`HTTP_SERVER` is used resulting in a path that begins with:
`/pub`.  A site that is stored in a sub-directory but does not
have a sub-directory in `HTTP_SERVER` will, however, begin with
the sub-directory defined by `DIR_WS_CATALOG` which would also
be known and technically could be removed from the full path as
well in the deconstruction of the `downloadsViaRedirect` class'
property `wsPubFolder`.