v156 Second of two solutions for download via redirect
This solution determines the sub-folder information from
`HTTP_SERVER` and prepends it to `DIR_WS_DOWNLOAD_PUBLIC`.
While it maintains a "server-less" characteristic allowing
the `zen_redirect` function to operate off of the domain name,
it is not easy to reverse engineer and actually requires a bit
of work to recreate even if a variable name were made available.

Recommendation (if it supports the proposed integration with other
code) is to use the first commit of this branch.