v157 - Always unset variable. (applies to v156 also, but...)
If the unsetting was on purpose (to provide a downstream
indicator about the success of decoding the session data), then
please ignore this commit (and please leave comment for
historical sake :)). Otherwise, it seems like the variable
`$backup` is not needed elsewhere after this assignment and
should be removed at each assignment.