Supports required file uploads.
As identified in the Zen Cart forum:
https://www.zen-cart.com/showthread.php?225489-Setting-up-a-product-with-a-single-required-file-upload-attribute
This offers a way to honor marking an File attribute as required.

This does not yet address the option name being identified as
Text when it is a file attribute.

Note, if two or more files are marked as required that this
will "upload" all files identified; however, if any required
file is not uploaded, then none of the uploaded files get
attached to a store product and would need to be uploaded again.
Not sure yet if/how to address this, but maybe another will be
able to help or "accept" that issue.