v157 PHP 7.3 notice of non-variable passed by reference
Fixed #2464.

Created a variable specifically to pass the sanitized information to the observer and then deleted the variable.  This maintains the notifier as previously provided while still providing sanitized information to the second parameter (editable by design of the observer class beginning in Zen Cart 1.5.3) but is then unset to not affect the subsequent storage of the original provided `$email_address`.