v156+ Add val_function database email information for new install
Adds the val_function field definition to perform checks on entry of email address to support sanitization as considered appropriate for an email address.  Entries do not discriminate at this time of requiring the email address to be of any of the following formats:

`name <email_address>`,
`email_address`,
`<email_address>`, or
combinations of the above separated by commas such as: `name1 <email_address1>, <email_address2>, name2 <email_address3>, email_address4`.

Where `email_address` with a number at the end or not is any valid email address and `name` with or without a number is the text to be displayed when the actual email address is to be hidden.