v157 paypalwpp.php zen_get_file_directory support template override
The first parameter to the function `zen_get_file_directory()` is expecting a path that ends with a slash to support joining the data presented.  If it is not provided there and it is not provided as the filename prefix, then the returned value may not be a valid filename, at the least it will not be the file expected/desired as the filename will be joined to the language.